UBR-135: Permanently fixed PagedCollection failure in ActivityControllerTest, but needs fixing in Hydra.
Conflicts:
	test/Activity/ActivityControllerTest.php
	test/Activity/data/activities.json

+++ test/Activity/ActivityControllerTest.php
@@ -197,7 +197,12 @@ class ActivityControllerTest extends \PHPUnit_Framework_TestCase
             ->expects($this->once())
             ->method('search')
             ->willReturn(
-                new PagedResultSet(new Integer(10), $activities)
+                new PagedResultSet(
+                    // The default limit is 5, so to have 3 pages of results we
+                    // need a total of at least 11 results.
+                    new Integer(15),
+                    $activities
+                )
             );
 
         $routeName = 'route-used-for-activities';
